Fixing warnings for Linux using GCC 4.8.3 (except in view_mensural.cpp)

+++ src/att.cpp
@@ -219,8 +219,8 @@ data_MEASUREBEAT Att::StrToMeasurebeat(std::string value) const
     }
     int measure = 0;
     double timePoint = 0.0;
-    size_t m = value.find_first_of('m');
-    size_t plus = value.find_last_of('+');
+    int m = (int)value.find_first_of('m');
+    int plus = (int)value.find_last_of('+');
     if (m != -1) measure = atoi(value.substr(0, m).c_str());
     if (plus != -1) {
         timePoint = atof(value.substr(plus).c_str());
